Regenerate Pipfile.lock on GNU/Linux to get pexpect
It should include pexpect as a dependency on Unix-like systems no
matter what platform the lockfile is created on, but it doesn't.

This fixes the problem that, with pipenv on Unix-like systems (or
at least on GNU/Linux), the notebooks could not be run. This
affected both JupyterLab and running them in VS Code.

Before merging this, it's necessary to test it on Windows to make
sure all needed dependencies on Windows still get installed.