Update some dependencies (#26)
But I'm keeping typeguard back at major version 2 for now, as
detailed below.

* Update dependencies

* Keep typeguard back so typechecked works in notebooks

This seems to have stopped working in major version 3 (based on
my testing in this project). The current major version of typeguard
is 4. The issue looks like:

https://github.com/agronholm/typeguard/issues/364

This should probably only be a temporary solution. There are other
libraries I could use. But also, it's not obvious that the added
clarity of presentation from defining the top-level functions in
the notebooks that use them, rather than in modules as I would
ordinarily do, is justified, even in the absence of this issue. So
it may be worthwhile to revisit that design decision for codegraph.
But for now I'm keeping that as it is and holding back typeguard.

This commit also updates typing-extensions to the next patch
version.

* Re-run notebooks